+++ javascript/static/index.mjs
@@ -452,6 +452,18 @@ export function stateToggle(element, token) {
 }
 
 /**
+ * See `stateAdd()`.
+ */
+export function stateContains(element, token) {
+  const state = new Set(
+    (element.getAttribute("state") ?? "")
+      .split(" ")
+      .filter((token) => token !== ""),
+  );
+  return state.has(token);
+}
+
+/**
  * > **Note:** This is a low-level function—in most cases you want to call `mount()` instead.
  *
  * Execute the functions defined by the `javascript="___"` attribute, which is set by [`@radically-straightforward/build`](https://github.com/radically-straightforward/radically-straightforward/tree/main/build) when extracting browser JavaScript. You must call this when you insert new elements in the DOM, for example, when mounting content.